reorganization components and custom react-routes-dom
