+++ package.json
@@ -81,9 +81,10 @@
     "prop-types": "^15.5.8",
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
-    "react-redux": "^4.0.0",
-    "react-router": "^4.1.1",
-    "redux": "^3.0.4",
+    "react-redux": "^4.4.4",
+    "react-router": "^2.0.1",
+    "react-router-dom": "^4.1.1",
+    "redux": "^3.4.0",
     "redux-devtools-extension": "^2.13.0",
     "redux-thunk": "^2.1.0",
     "sass-loader": "^6.0.3",
