+++ src/js/components/main-nav/index.jsx
@@ -1,0 +1,17 @@
+/**
+ * Created by Denis on 09.06.2017.
+ */
+import React, { Component } from 'react';
+import { Link } from 'react-router-dom';
+
+export default class MainNav extends Component {
+  render() {
+    return (
+      <ul>
+        <li><Link to="/">Airports</Link></li>
+        <li><Link to="/arrival">Arrival</Link></li>
+        <li><Link to="/departure">Departure</Link></li>
+      </ul>
+    );
+  }
+}
