+++ src/js/containers/departure.jsx
@@ -3,12 +3,16 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 // модуль комбинирования работы нескольких актшионов
 import { bindActionCreators } from 'redux';
+import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import getMuiTheme from 'material-ui/styles/getMuiTheme';
+import { deepOrange500 } from 'material-ui/styles/colors';
 import * as getActionData from '../redux/actions/actions';
 // подгрузка компонентов
 import SheduleDeparture from '../components/shedule-departure/';
+import MainNav from '../components/main-nav';
 
 
-class Departure extends Component {
+class ShedulesDeparture extends Component {
   static get propTypes() {
     return {
       getData: PropTypes.object.isRequired,
@@ -20,9 +24,17 @@ class Departure extends Component {
   }
   render() {
     const departing = this.props.currentStore.Reducer.shedules.departing;
+    const muiTheme = getMuiTheme({
+      palette: {
+        accent1Color: deepOrange500,
+      },
+    });
     return (
       <div>
-        <SheduleDeparture infoData={departing} />
+        <MainNav />
+        <MuiThemeProvider muiTheme={muiTheme}>
+          <SheduleDeparture infoData={departing} />
+        </MuiThemeProvider>
       </div>
     );
   }
@@ -35,4 +47,4 @@ const getDataProps = dispatch => (
 export default connect(
   state => ({ currentStore: state }),
   getDataProps
-)(Departure);
+)(ShedulesDeparture);
